+++ src/main/java/com/scorpiac/javarant/DevRantAuth.java
@@ -21,12 +21,12 @@ public class DevRantAuth {
      * @param vote The vote to cast.
      * @return The rant.
      */
-    public Result<Rant> voteRant(int id, Vote vote) {
+    public Rant voteRant(int id, Vote vote) {
         return devRant.getRequestHandler().post(
                 ApiEndpoint.RANTS.toString() + '/' + id + '/' + ApiEndpoint.VOTE.toString(),
                 RantResponse.class,
                 getParameters(vote.getOptions())
-        );
+        ).getValueOrError();
     }
 
     /**
@@ -36,19 +36,19 @@ public class DevRantAuth {
      * @param vote The vote to cast.
      * @return The comment.
      */
-    public Result<Comment> voteComment(int id, Vote vote) {
+    public Comment voteComment(int id, Vote vote) {
         return devRant.getRequestHandler().post(
                 ApiEndpoint.COMMENTS.toString() + '/' + id + '/' + ApiEndpoint.VOTE.toString(),
                 CommentResponse.class,
                 getParameters(vote.getOptions())
-        );
+        ).getValueOrError();
     }
 
     private NameValuePair[] getParameters(List<NameValuePair> params) {
         // Add the auth parameters.
-        params.add(new BasicNameValuePair("token_id", devRant.auth.getId()));
-        params.add(new BasicNameValuePair("token_key", devRant.auth.getKey()));
-        params.add(new BasicNameValuePair("user_id", devRant.auth.getUserId()));
+        params.add(new BasicNameValuePair("token_id", devRant.getAuthObject().getId()));
+        params.add(new BasicNameValuePair("token_key", devRant.getAuthObject().getKey()));
+        params.add(new BasicNameValuePair("user_id", devRant.getAuthObject().getUserId()));
 
         return params.toArray(new NameValuePair[0]);
     }