+++ src/main/java/com/scorpiac/javarant/NoSuchUserException.java
@@ -1,0 +1,7 @@
+package com.scorpiac.javarant;
+
+public abstract class NoSuchUserException extends RuntimeException {
+    public NoSuchUserException(String message) {
+        super(message);
+    }
+}