+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.1.0
+
+Add an exception model and change the return values of all API methods.
+All methods now simply return the value, or throw an exception if something went wrong.
+
 ## 2.0.0
 
-Initial version (since rework).
+Initial version.